In the process of fixing the pelicanconf.py file.
Cleaned it up real good and added some comments. Need to go through the documentation and add all the settings that I care about. This has made the development process a bit simpler. I might do some other stuff to change the backend structure of how the pages are created.